refactor(home): refactor home section code
refactors the code in the home section
