+++ src/App.js
@@ -1,6 +1,6 @@
 import GlobalStyles from "./components/styles/Global";
 import Nav from "./components/Nav/Nav.js";
-import Home from "./components/Home";
+import Home from "./components/Home/Home";
 import Projects from "./components/Projects/Projects";
 import About from "./components/About";
 import Contact from "./components/Contact";
