+++ src/components/Home/Variants.js
@@ -1,0 +1,27 @@
+export const greetingVariant = {
+  hidden: {
+    y: "20vh",
+    opacity: 0,
+  },
+  visible: {
+    y: 0,
+    opacity: 1,
+    transition: {
+      duration: 1.25,
+    },
+  },
+};
+
+export const heroVariant = {
+  hidden: {
+    y: "-20vh",
+    opacity: 0,
+  },
+  visible: {
+    y: 0,
+    opacity: 1,
+    transition: {
+      duration: 1.25,
+    },
+  },
+};
