--- src/components/styles/Home.styled.js
@@ -1,34 +1,0 @@
-import styled from "styled-components";
-
-export const StyledHome = styled.div`
-  display: flex;
-  align-content: center;
-  justify-content: center;
-  align-items: center;
-  height: 90vh;
-  margin-left: 100px;
-
-  .hello {
-    display: flex;
-  }
-
-  h1,
-  h5 {
-    color: #fefefe;
-    margin: 0 0;
-    padding: 0 0;
-  }
-
-  h1 {
-    font-size: 80px;
-  }
-
-  h5 {
-    font-size: 21px;
-  }
-
-  img {
-    margin-left: 80px;
-    height: 480px;
-  }
-`;
